content/html/content/test/test_bug546995-2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=546995
-->
<head>
  <title>Test for Bug 546995</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=546995">Mozilla Bug 546995</a>
<p id="display"></p>
<div id='content'>
</div>
<pre id="test">
<script type="application/javascript;version=1.8">

/** Test for Bug 546995 **/

SimpleTest.waitForExplicitFinish();

var gGen = runTests();
addLoadEvent(function() { gGen.next(); });

var gElement = "";

function runTests()
{
  // TODO: keygen should be added when correctly implemented, see bug 101019.
  var elements = ["input", "textarea", "select", "button"];

  for each(element in elements) {
    var iframe = document.createElement("iframe");
    document.getElementById('content').appendChild(iframe);
    var iframeCw = iframe.contentWindow;

    gElement = element;
    iframeCw.location = "file_bug546995.html";
    yield;
    is(iframe.contentDocument.activeElement,
       iframe.contentDocument.getElementById('a'),
       "The first inserted element with autofocus should be focused");

    document.getElementById('content').removeChild(iframe);
  }

  // Now we want to focus the body element.
  document.activeElement.blur();
  is(document.activeElement, document.body,
     "Body should be the active element");

  // Adding elements with autofocus.
  for each(element in elements) {
    var e = document.createElement(element);
    e.autofocus = true;
    document.getElementById('content').appendChild(e);
  }

  // When the function will be executed, the focus event from autofocus should
  // have been processed.
  SimpleTest.executeSoon(function() {
    ok(document.activeElement, document.body,
       "After loading, elements can't be autofocused");
    SimpleTest.finish();
    });
  yield;
}

</script>
</pre>
</body>
</html>
